+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.73] - 2019-12-25
+- [Add Paredit drag up/down commands](https://github.com/BetterThanTomorrow/calva/issues/500)
+- [Add Paredit drag forward up/backward down commands](https://github.com/BetterThanTomorrow/calva/issues/500)
+
 ## [2.0.72] - 2019-12-13
 - [Deselect text after surrounding with parens/braces/etc](https://github.com/BetterThanTomorrow/calva/issues/511)
 - Fix: [Strict mode backspace/delete not deleting unbalanced brackets](https://github.com/BetterThanTomorrow/calva/issues/501)